// app/_globalRedux/services/types/alerts.d.ts
export interface AlertByDayEvents {
	starttime: string;
	endtime: string;
	vehicle_no: number | string;
	exception_type: string;
	KM: string;
	duration: string;
	startlocation: string;
	startlat: number;
	startLong: number;
	endlocation: string;
	endlat: number;
	endLong: number;
	speed: number;
	journey_statusfinal: string | null;
	Halting?: string | null;
	hour?: string;
	InvoiceNo?: string;
	InvoiceDate?: string;
	remark?: string;
	id: number;
	service_id: string;
	route_name: string;
}